method __str__ from UserAgent now using methods (#94)
Is better to calculate and provide a method to calculate the string values for `device`, `os` and `browser`  instead of doing it  directly on `__str__` method. 

This way the users may use those methods to get each data directly. Instead of having to make a `user_agent.__str__().split('/')` to get each value